requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.8
multidict==6.1.0
phonenumbers==8.13.45
pillow==10.4.0
pycountry==24.6.1
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
twilio==9.3.1
tzdata==2024.1
urllib3==2.2.2
yarl==1.12.1
razorpay==1.4.2
setuptools==75.2.0
#cffi-1.17.1 cryptography-43.0.1 pycparser-2.22